moved logic to controller layer
